test/t0001-basic.t: speed up test
Problem: t0001-basic.t takes a long time to run.

This test includes many back to back instance invocations,
which due to #1006, each run out the default "shutdown-grace"
timeout, which for size < 16 is 1.0s.

Speed things up by shortening the grace period to 0.1s.
This means module unloading in rc3 will certainly be interrupted,
however, that should not affect the outcome of any of the tests
in this script.