modules/connector-local: avoid unload deadlock
Problem: if connector-local module is unloaded while a client is
still connected that has active event subscriptions, the
connector-local teardown will call flux_event_unsusbscribe(),
but in most cases connector-local is torn down as a special
case after the broker reactor has terminated, so this will deadlock.

Avoid calling flux_event_unsusbscribe() when this module is
unloaded.  Its subscriptions are part of the module context that
will be destroyed when the module unloads anyway, thus unsubscribing
to all subscribed-to events serves more of an aesthetic purpose
than a practical one.

Fixes #1025