libflux/message: add FLUX_MSGFLAG_PRIVATE
Add FLUX_MSGFLAG_PRIVATE message flag, and accessors:

flux_msg_set_private()
flux_msg_is_private()

This bit can be set on any message.  If set, connectors
must not forward private messages to connections other
than those authenticated with the instance owner role,
or the same userid as the sender of the message.

The motivating use case is to protect the kvs.setroot
event which exposes the KVS root blobref.  Direct use of
the KVS is restricted to the instance owner, but we would
like to make the content store available to guest users.