modules/kvs: set privacy bit on kvs.setroot
As a precursor to opening content-cache.load to guest
users, make the kvs.setroot (and kvs.error) events private
to the instance owner, as these are intended to be
"internal" to the KVS service.

A guest user should not be able to subscribe to kvs.setroot
and obtain the root blobref, which could would allow the
guest to walk the entire KVS namespace.