broker/content-cache: open load/store to USER role
Allow users with FLUX_ROLE_USER to load/store blobs from/to
the content cache.

While the KVS service is only available to the instance owner,
the content cache, which serves as backing store for KVS objects,
can be safely made available to guest users as long as the
blobrefs that refer to KVS objects are not exposed to guest
users.