systemd: set rundir to $runstatedir in flux.service
For systemd unit file, use the value of runstatedir for the
broker.rundir attribute, instead of using systemd's idea of the
runtime directory with `%t`.

This ensures that the "default" FLUX_URI calculated by the installed
flux binary agrees with the actual flux session started by systemd.

This also allows *both* the rundir used for the systemd initiated
flux instance, and the one used for the default URI calculation to
be set by a single option at configure time, thus they will never
be out of sync., e.g.

 ./configure --runstatedir=/run ...

for a more standard systemd install.