build: don't install systemd unit file outside of prefix
When --with-systemdsystemunitdir is not set explicitly, prepend
the pkg-config systemdsystemunitdir with ${prefix}, so that the
unit file is not installed outside of --prefix.

Distro packages should therefore use --with-systemdsystemunitdir
explicitly, in the case where the unit file directory *is* outside
of prefix (e.g. /lib/systemd/system when --prefix=/usr).