cmd/flux-list-instances: drop command
Problem: flux-list-instances no longer works, has no test
coverage, and uses a fairly ad-hoc and fragile method
to find instances.

As discussed in #1034, this command probably should be
replaced with a scheme that uses the system instance
as the point of contact.

Drop the command and its helper function over in the
python bindings.