broker/namespace: New service
Add a new broker-resident mini-service for mapping a fixed name
to a moveable content reference, similar to the KVS setroot/
getroot operations.  It implements the operations:

  ns.create  - create a new namespace
  ns.remove  - remove a namespace
  ns.commit  - update a namespace
  ns.lookup  - get the current version of a namespace

The namespace state consists of the name, flags, owner, a sequence
number, and an untyped JSON object, intended to be a content store
reference such as a RFC 11 treeref, legacy KVS dirent, or a blobref;
however the application can use whatever it wants.

The sequence number must monotonically increase on each update.

If it is created with the FLUX_NS_SYNCHRONIZE flag, then each
operation also generates an event which causes the operation to be
mirrored on all broker ranks.  These events are:

  ns.allcreate       - create a new namespace
  ns.allremove       - remove a namespace
  ns.allcommit.NAME  - update a namespace NAME

Multi-user support:

The create and remove operations are reserved for the instance
owner.  A userid parameter is provided as part of the create request,
to allow the instance owner to create namespaces that are private to
a specific user.  The commit and lookup operations either must come
from the instance owner or from the namespace owner (else EPERM).

Internal events that accomplish the loose synchronization between
ranks are private to the instance owner and are ignored if sent by
another userid.  The internal ns.allcommit.NAME event is private to
the namespace owner and instance owner, and may be subscribed to by
those users to track their namespace updates.