libflux/message: add flux_msg_aux_get/set
Allow aux data to be associated with a message object.

The main use case for this is to allow message parsing
functions to return memory "owned" by the message, that will
be destroyed when the message is destroyed, so the user
doesn't have to manage that allocation separately.

Since message handlers receive a (const flux_msg_t *) parameter,
the design of flux_msg_aux_set() rationalizes taking a const
message and modifying it anyway with the argument that this
is an "annotation" that doesn't actually change the content
of the message.