libflux/reactor: avoid use of xzmalloc and oom
xzmalloc() calls oom() on failure which triggers and exit of
the calling program. This should not be done in libflux context,
ro replace the calls to xzmalloc() with calloc() and manually
check return codes.

Instead of repeatedly calling the same sequence of error handling
when flux_watcher_create() fails (free impl, set errno, return), a
convenience function was added `flux_watcher_create_cleanup`,
which handles cleanup, so that a single return can be used on
failure.