libflux: don't nullify argument to flux_handle_destroy
In the ssh connector, the flux handle is saved as a member of the ssh
connector implementation object. On an error path during initialization,
the cleanup calls flux_handle_destroy as

 flux_handle_destroy (&c->h);

where `c` is the connector object.

The implementation destroy operation (h->ops->impl_destroy()) is called
during flux_handle_destroy(), which frees the memory for `c`. As a final
step, flux_handle_destroy() attempts to nullify its argument (&c->h), which
now points to invalid memory and will possibly cause memory corruption and
a crash.

There are multiple ways to fix this, but the most straightforward is
probably to change the prototype of flux_handle_destroy() to take a
reference to `flux_t` instead of a pointer to a reference, and remove
the attempted nullification of its argument.

Fixes #1062