modules/kvs: Refactor load() logic
The load() function may load data from the KVS cache or it may
issue a rpc to retrieve data from the broker content cache if the
data is not yet locally cached.

The load() function's dependency on both cache and rpc APIs makes
it difficult to splice out functions that call load() into separate
APIs.

This patch removes the call of load() from the lookup() and
walk() functions.  In lookup() and walk(), data is instead looked
up only in the local KVS cache.  If it is not stored there locally,
it is the responsibility of the original caller to then call load()
to load the missing data.  A in/out parameter will pass the missing
reference all the way back to the original caller.