libflux/rpc: drop internal dead code
With flux_rpc_get_nodeid() removed, code that caches the
nodeid for singleton rpcs, and that inserts the nodeid
in the group matchtag for group rpcs, can be removed
to simplify the code.

Also, code for conditionally requesting a group matchtag
if more than one response is expected can be removed
as now only zero or one response can ever be expected.