libflux/rpc: simplify rpc state
It's no longer necessary to keep a tally of expected versus
received messages.  Expected is always zero or one
(as determined by flags), and received  is always zero or one
(as determined by whether rx_msg is set NULL or not.

Try clarify some convoluted logic in the continuation
registration.