libflux/reactor: don't copy watcher ops into each instance
Make the 'ops' vtable member of flux_watcher into a pointer to an
watcher_ops structure, instead of copying the vtable into each instance
of the watcher class.

This then obviates the need for watcher "signatures", because the
address of the operations structure can be used to uniquely identify
each "class" of watcher.

Besides, since the signatures were defined inside the source of
reactor.c, there was no way to guarauntee that any externally created
watchers would have a different signature from existing watchers;