connectors/shmem: add bind capability
Add bind capability to the shmem connector.

If user calls flux_open ("shmem://uuid&bind") then the socket
is bound instead of connected.  This allows two shmem connectors
to be used back to back in testing.

For symmetry also parse "shmem://uuid&connect".  This is the default
if no options are provided.