libflux/dispatch: tighten fastpath response match
Problem: dispatch fastpath does not check if a response is
in the correct "matchtag domain", so a flux module performing
a routing function like the connector-local module, might
inadvertantly confuse one of its RPC responses for one intended
for a connected user.

The only RPC that I know of that fits this profile is the
subscribe/unsubscribe RPCs made from connector-local to the
broker.  With changes to flux_rpc() to use futures, the
matching code switched from flux_recv() with a struct flux_match
to the dispatcher, so while this is an old bug, its manifestantion
is new with futures.