libflux/rpc: reimplement with futures
Change the RPC API to return flux_future_t instead of
flux_rpc_t.

Leave compatability in place by making flux_rpc_t an alias
for flux_future_t and aux_get/set, rpc_then, and rpc_check
there as wrappers for similar future calls.

When handling the response, call flux_response_decode()
on it and fulfill the future with an error if there was one.
This allows no-payload responses to be checked with
flux_future_get().

Fixes #1088