test/rpc: avoid flux_rpc_check() busy-wait
Stop using idiom in tests where we call flux_rpc_check()
in a tight loop until a result is ready.  We have changed
flux_rpc_wait_for() so that a wait with timeout=0 does not
invoke the reactor, so the idiom won't work.  And if it did,
it would be something to discourage due to its inefficiency.