libflux/barrier: fix inconsistent handle state key
Problem: flux_barrier(), when called with a NULL name,
creates a unique barrier name based on a counter which
it stores in the flux_t handle under a key.  However,
the keys used to store and retrieve this state are
inconsistent so it starts from scratch each time, resulting
in non-unique barrier names.

Make the aux key names consistent.