libkvs: move from kvs module directory
Relocate the KVS API from src/modules/kvs to src/common/libkvs.
This is built as a static library libkvs.la which is included
in the installed dynamic library libflux-core.so.  Temporarily
unmask symbols with kvs_*, kvsdir_*, and kvsitr_* prefixes
in that library.

Include proto.[ch] and json_dirent.[ch] in libkvs.la.
Since these are used by the KVS module also, but are not
exported via the public libflux-core.so, the KVS module
links statically with libkvs.la to obtain access.

Various Makefile.am and include paths are updated.