modules/kvs: Refactor commit_unroll()
Instead of calling store(), which will store to cache and
send rpcs to the content store, instead only store to cache
and return a list of entries which need to be sent to content
store.  Caller to commit_unroll() is responsible for sending
data to content store and waiting on dirty entries.

Remove store(), as it is no longer used.