t/t5000-valgrind.t: disable unless valgrind.h found
Disable t5000-valgrind.t by default when valgrind/valgrind.h was
not found by ./configure. This means that the valgrind hook to
disable dlclose() for modules is not active, and this has been found
to cause false positives for this test.

The test can still be forced by hand with the use of -d, --debug flag,
e.g.

 ./t5000-valgrind.t -d

Fixes #1097