test/t1002-kvs-cmd.t: cleanup
This test runs --verbose, these errors are omitted:

./t1002-kvs-cmd.t: 352: [: 0: unexpected operator
./t1002-kvs-cmd.t: 352: [: 0: unexpected operator
./t1002-kvs-cmd.t: 380: [: 1: unexpected operator

Use [ $i -eq 50 ] to compare numbers rather
than [ $i == "50" ].  I'm not sure what, if any, effect
this actually had on the tests.  I seemed to have some
intermediate failures that went away after fixing this,
so possibly the while loop was exiting on the first
iteration, and that worked most of the time.

Also, tidy up here documents by replacing << with <<-
and indending them.