libkvs/watch: convert to jansson
Convert kvs_watch() internals from json-c to jansson.

This conversion may have introduced a bug that causes
t/kvs/watch simulwatch with ntimes > 499 to hang.

For code simplification, one new thing here is that each
watch is assigned a future.  After the first response, the
future is fulfilled, but is kept around to prevent the matchtag
from being freed.  This simplified the code for performing the
initial RPC, but does increase the amount of state kept per
watch.  We really need a more well thought out way to handle
the "one request - multiple response" use case.

After the first response that fulfills the future, a generic
response message handler is used for subsequent responses.