test/kvs: reduce number of simultaneous watches
t1000-kvs-basic.t calls
  t/kvs/watch simulwatch <key> 8192

This is a stress test of sorts for the kvs_watch client
code.  Drop the count from 8192 to 256 to avoid a future problem
at exactly 500 (not 499) concurrent watches of the same key
where the test hangs.  The problem begins occurring after
the jansson rework of kvs_watch - modify the test here to
avoid breaking git bisect.

This may be an actual bug, but since kvs_watch() has significant
other changes pending, and many simulateous watches is not a real
use case anyway, leave it for another time.