test/python: drop kvs test_set_none() check
This test appears to be wrong - putting a null key is
equivalent to an unlink, but the test fails when that
actually happens.  The test may have been relying on
an earlier bug where the value of a key became the JSON
string "null" instead of the json value null.

Drop the test for now.