libkvs/jansson_dirent: reject FILEVAL with JSON null
Careful: json-c and jansson handle JSON null differently:

json-c decodes "null" as a NULL json_object *.
json-c encodes a NULL json_object as "null".
jansson encodes "null" as a valid json_t *.
jansson decodes a NULL json_t as an error.

Historically, kvs_put_obj () of a NULL json_object value was
equivalent to a kvs_unlink().  Due to the way json-c encodes
a NULL json_object as "null", this was special cased in the
kvs_put_obj() code, and thus a kvs key could never have
a JSON null value.

In addition due to the way the json-c dirent_validate() was
structured, a FILEVAL with a null value would not be matched
in the FILEVAL conditional leg and thus would fail the validity
check.

Ensure this remains the case through the conversion by updating
the jansson j_dirent_validate() to fail the validity check
if the value is JSON_NULL.

If we want to allow JSON null values in the KVS we could
do that, but we should make that change purposefully.