test/kvs: put key=null and ensure valid dirent
A flux_kvs_txn_put (key, "null") should result in an EINVAL.
The t/kvs/basic put subcommand retries an EINVAL with the value
converted to a string.

Add a test that ensures the value is set to a string rather
than a JSON null, which is an invalid directory entry.