bindings/lua: take a reference on watch directory
Problem: valgrind shows access to freed memory when
kvsdir object created by watch callback is garbage
collected.

When a kvs_watch() callback receives a kvsdir_t
argument, it is only valid for the duration of the
callback.  When instantiating a directory as a Lua table
that persists after the callback, it looks like we need
to call kvsdir_incref() on the directory to avoid illegal
access and/or double free during garbage collection.

Fixes #1112