modules/kvs: Fix dirty cache logic
In earlier versions of the KVS, code logic could allow
cache_entry_wait_notdirty() to be called multiple times
on the same cache entry.  By itself, introduced in
0cbeacc09aba444f2c972cd7cd83846d721922f7, this was harmless.

This code logic was carried over spiritually into the
commit API and its callback commit_iter_dirty_cache_entries().
This required the addition of flags in the cache to determine
if a cache entry had been flushed to the content store yet or not.

With the commit API and its dirty cache list, it is not longer
necessary nor reasonable that a cache entry be added to the list
multiple times.  Changing this logic will clean up the code and allow
future error handling corner cases to be handled more easily.

Adjust commit unit tests accordingly.