flux-sharness.sh: add NO_CHAIN_LINT prereq
Some sharness tests are not able to  make --chain-lint happy, so add a prereq
`NO_CHAIN_LINT` so that these tests are skipped during runs with --chain-lint,
and the explicit broken &&-chain is explicitly purposeful.

The main case where &&-chains are broken and it is ok is when a process
is sent into the background during a test.