cmd/flux.c: canonicalize bindir in flux_is_installed
The check for whether the flux command driver is installed or in
the build tree does not canonicalize X_BINDIR, so if this path
includes a symlink the test will incorrectly fail.

Canonicalize the path using `realpath(3)` before the comparison
to avoid missing detection of installed flux. If realpath() returns
ENOENT, then the install path doesn't exist and we must be running
from the builddir.

Fixes #1121