modules/kvs: Remove oom() from wait_runqueue()
Re-work wait_runqueue() to not oom() on error.  Algorithm within the
function has been altered to be safe against errors which requires
that all elements on the waitqueue be copied off or none.  This change
effectively converts the function from having two loops to having
three.

Update callers to check for error.  Notably adjust cache_entry_set_dirty()
and cache_entry_set_json() to return -1 on error, and update callers of
these functions.

In some locations, where an errno had to be set to indicate a missing
cache entry, errno = ENOTRECOVERABLE was chosen.  In another location where
a cache entry was missing data, errno = ENODATA was chosen.