libutil/cronodate: add cronodate_set_integer()
Add a function to set a single integer value for any time unit
member of a cronodate object. This is a convenience function to
prevent requiring a single integer to be converted to a string
and passed in as a range, which is then converted right back to
an integer.