modules/cron: avoid implicit rounding of timestamps
Avoid implicit rounding in timespec_to_double in both cron.c and task.c.

For the cron task timestamps, rename the local function
round_timespec_to_double() to be honest about what we're doing,
and expand the comment to explain why.

For other cron timestamps, just drop the rounding for now. If it
an inconvenience in the future (unlikely), the rounding can be
copied from task.c as an example.

Resolves #1131