libflux/request: Fix "data" type flux_request_decode_raw()
Fix the type of the "data" parameter in the flux_request_decode_raw().
In all uses, "data" is a the address of a pointer, so "void **data" is
the correct type.  This makes it fairly clear that the buffer will be
provided by flux_request_decode_raw() rather than by the caller.