broker/service: allow multiple service names
Instead of only allowing one service name with an optional alias,
allow any number of services to be registered with the same callback.
Pass in a uuid with each registration and add a function to remove
all services registered by a given uuid.

Modify the broker so that modules register their primary and
alias service names with the module uuid, and remove all services
by uuid when the module unloads.

Optimize service_send() a bit so that a malloc is not incurred
each time we want to lookup a service by a substring of a message's
topic string.  (If it's short, use a stack variable).

Fix up remaining spots where exit is called on malloc failure.