t/kvs: Fix error in null tests
In original test of setting the string "null" within the KVS there
was an error that worked out by chance but was incorrect.

When doing a kvs put with the string "null", such as with:

flux kvs put key=null

this should create a json null object.  On the other hand:

flux kvs put key=\"null\"

should create a string "null".

Correct for this and create tests for both scenarios.