broker: only register single-word service names
Problem: we inconsistently register service names for
individual request handlers in the broker, with the same
callback.  This complicates the service lookup code
(since it has to try to lookup multiple substrings of inbound
message topic strings).  It's also unnecessary.

Clean this up so we only register single word services.