flux-sharness.sh: detect failure to unload modules
Detect failure to unload modules when test_under_flux is used
and modules are loaded manually during testing. Failure to unload
modules can cause lack of coverage and is considered an error.

Unfortunately, the only practical way to generate an error is
to add a new "cleanup" hook, and fail during this hook with an
"exit 1" if any modules are loaded that were not also loaded
initially (these will presumably be unloaded during rc3).

Fixes #1020