modules/kvs: don't complain of invalid root cache entry
Problem: the heartbeat event handler tries to "touch"
the root cache entry to prevent it from being temporally
expired, for performance.  It calls load() on the root
blobref, but load() complains if there happens to be
an invalid (content load pending) cache entry for the root.
Severe sounding errors are emitted:

  kvs.err[2]: synchronous load(), invalid cache entry
  kvs.err[2]: heartbeat_cb: load: No data available

Call cache_lookup() on the root blobref instead, and don't
complain if the lookup fails.

Fixes #1174