modules/kvs: [cleanup] setroot_event_send() use cache
Problem: setroot_event_send() calls load() on the root cache
entry, to look up the root dir object for inclusion in the
setroot event payload.  However, most of the code in load()
is not used here since the the root dir object is always in
cache on rank 0.

Simplify the code by calling cache_lookup() instead.