subprocess: avoid enum in subprocess_manager_set
Change prototype of subprocess_manager_set() to avoid using an enum
as the argument passed to va_start(), as this causes warnings on newer
compilers which complain about

      passing an object that undergoes default argument promotion
      to 'va_start' has undefined behavior [-Werror,-Wvarargs]

The enum undergoes default promotion, and thus it may cause undefined
behavior. Therefore, change the enum argument to an integer to avoid
this warning.