libflux/future: drop reactor arg from init callback
Drop the reactor argument from the future's initialization
callback.  It should be accessed in the same manner as
the flux_t handle now that we haver reactor get/set functions.
This makes the API more uniform, and the reactor is often not
used if the future contains message handlers instead of
reactor watchers.

Update rpc implementation and unit test.