modules/kvs/lookup: Update for raw opaque data
Update internal lookup API to handle valref references pointing to
raw unencoded data in the content store.

Most notably, missing references have an additional boolean indicating
if data is raw, so that on data load it can be handled appropriately.

In addition, when returning data pointed to by a valref, add checks
to ensure the data is raw in the KVS cache.

Adjust lookup unit tests appropriately.