job-ingest: Introduce a job-ingest prototype
Introduce a new module named "job-ingest".  It is in early prototype stage.
It can accept a jobspec submission message and run the jobspec through
libjobspec for basic parsing validation.  It replies with error if the jobspec
fails to parse, and returns a string containing a job number if it parses
successfully.

Introduce the new command "flux-jobspec-submit".  It can take one or more
jobspec files on the command line, or accept a jobspec stream from stdin.
Each jobspec is individually packaged into a request for the job-ingest module.