flux.service: Fix EXTRA_DIST and .gitignore
Add flux.service .gitignore, since it is generated from
flux.service.in at configure time.

Remove the EXTRA_DIST of flux.service. Since this is generated
at configure time from flux.service.in and flux.service.in
is already part of the dist by way of being declared in AC_CONFIG_FILES,
it does not appear that flux.service should be EXTRA_DIST.