build: Convert @GCCWARN@ to $(WARNING_CFLAGS)
Change the name of GCCWARN to WARNING_CFLAGS to eliminate
the gcc-specificity from the name.  Granted it is unlikely, but
WARNING_CFLAGS might potentially have different values depending
on the detected compiler, without requiring changes to each
Makefile.am.  The name is also now in keeping with our other
*_CFLAGS named variables.

Also, we change from @GCCWARN@ to the more common variable usage
$(WARNING_CFLAGS).

This will help a little bit with coming C++ flags.