libkz: modernize KVS API usage
Modernize some aspects of libkz's KVS API usage.

I held back converting the "anonymous" transactions
to explicit ones, as doing so seemed to hang up the
wreck tests.  My best guess is that something in wrexecd
that is written to the anonymous transaction is actually
being committed by an anonymous commit/fence in libkz.
Since this code is due to be replaced it didn't seem like
time well spent right now to run this down.