modules/kvs: Refactor lookup handling of valrefs
Refactor internal lookup API to be able to handle returning multiple
missing references from a valref to the caller.  This is predominantly
infrastructure support for future multiple missing references support
and nothing in the internal KVS lookup API currently uses this.

As a fallout of this refactoring, the `missing_ref_raw` flag is no
longer necessary and has been removed.