module/kvs: Modify cache_entry_create() style
cache_entry_create() now takes a cache_data_type_t parameter,
allowing to user to create a cache entry without any data
attached to it, but at the same time indicate the type of data
expected for the cache entry to store.  This is an alternate
to passing NULL to the function cache_entry_create_json ().