modules/kvs: Add cache internal data_valid flag
In the internal cache API, add a data_valid flag which indicates
if the user has set data or not.  This flag is used instead of
checking data == NULL for validity, as checking the data pointer
for NULL would not allow zero length content-blobs to be stored
in the cache.

A major fallout of from this addition is the change to the
cache_entry_set_raw() function.  It was previously unclear
if passing NULL & length 0 to cache_entry_set_raw() indicated
if you were setting a zero length data or trying to clear
prior data.  Now it is made clear that this is setting zero
length data and clearing data is through cache_entry_clear_data().
In addition, calling cache_entry_set_raw() with zero length
data can result in wait_runqueue() being run on hp->waitlist_valid.

Added new unit tests for coverage.