broker/attr: Add attr_add_[int|uint32] functions
There isn't much need to maintain references to external int and
uint32_t values when those values are marked immutable.  On the contrary,
unless those eternal values are very carefully document we introduce
an opportunity for accidental errors.

Furthermore, it is simpler and more efficient to perform the conversion
into a C string just once rather than implementing a callback that
performs the conversion on every access.

We add these new wrappers to enable cleanup elsewhere in the code.