broker: Move rank and size out of broker_ctx_t
Relocate rank and size from broker_ctx_t into overlay_t.
This keeps the information about the size of the overlay network
and our rank in the overlay network closer to the thing that
would know about and manage the overlay network.

Eventually more will be abstracted away into the overlay portion
of the code, which will make it easier to offer alternate methods
of gathering information to feed to the overlay initialization
code.