libkvs/dir: skip re-encode in flux_kvsdir_tostring()
flux_kvsdir_tostring() calls json_dumps () on the encoded
JSON object stored inside the object.  Instead of doing this,
just save the original json_str passed to flux_kvsdir_create().

This also fixes a memory leak since the cached json_dumps()
result was not freed when the object was freed.