libkvs/txn: add flags to commit operation object
As discussed in #1193, the operation object (an array
element in the ordered array of operations sent with
commit/fence requests) needs to expand to support a
future KVS append operation.

Add 'flags', an integer value, to each operation,
always 0 in this commit.

Request parsing on the KVS service end currently ignores
the extra 'flags' value.