libkvs/txn: drop special case for txn_put of NULL
If flux_kvs_txn_put() is handed a NULL value, it calls
flux_kvs_txn_unlink().  This is a bit counter-intuitive now
that, for example, a zero-length value is valid.

Drop the special case from flux_kvs_txn_put(), and make
a NULL value an EINVAL error in this commit.

Add the special case into the deprecated functions
flux_kvs_put() and flux_kvsdir_put() so that lua bindings
continue to work.

Update tests.