cmd/flux-kvs: handle non-JSON values in "dir" cmd
Problem: the KVS allows non-JSON values to be stored, but
flux kvs dir -R exits with an error if it encounters one.

For each value, if it decodes as JSON, format it like before.
But if it doesn't, print it directly instead of exiting.
If the output is a terminal, truncate long values so they fit
within the terminal width (reusing logic from the ls subcommand).
Also truncate values where they contain unprintable characters
or newlines.  Indicate truncated values by appending "...".

Fixes #1158