cmd/flux-kvs: get --json handles zero-length value
Problem: if value is zero-length, get --json displays
"null", but it should be an error like any other value
that doesn't decode as valid JSON, for example the empty
string.

output_key_json_string() is doing double duty with the
watch subcommand, which still treats a NULL value as a
non-existent key.  Don't pass a NULL value to that function;
instead, treat it as an error beforehand.