libkvs/treeobj: decode_val add xtra 0 byte to data
Problem: if string values are not stored in the KVS with
their NULL terminators, we need to recopy them to a len + 1
size buffer before returning to the user.

base64_decode_length() already includes provision for an
extra NULL byte.  Add some assertions to the code to
ensure (and document) that this NULL byte is present,
then document the behavior in the header.

Add a unit test.