libkvs: do not store strings with \0 term
Problem: the reasons no longer exist for storing
string/json values in the KVS with their terminating
NULL byte, and this complicates the implementation
of atomic append.

Skip storing the NULL byte, and change lookup_get()
et al to rely on the fact that treeobj_decode_val()
pads the space after the returned data with a NULL byte.

Use json_loadb() instead of _loads() when decoding
a JSON value.

Update unit and sharness tests.

Fix #1261