libkvs/treeobj: use const on json_t where approp
Add a new internal function treeobj_peek() which accepts
a const json_t * input parameter, and produces a const json_t
*data output parameter.

Internally, treeobj_peek() calls json_unpack(), which required
the const on the input parameter to be cast away - safe since,
while we use the "o" format specifier to extract the data object,
we put a const on it before we return it to internal callers.

The treeobj_peek() function allows many functions to accept
const json_t * where before they did not.

There was one other place where a const had to be cast away.  In
treeobj_validate(), which now calls treeobj_peek() to access the
const data member, the data const had to be dropped while iterating
over its directory entries with json_object_foreach(), since
json_object_iter() does not accept a const object.  I'm hopeful
that the iterator does not actually modify the object as it seems
self-contained, and this is just an oversight in the jansson API design.

As discussed in #1264