modules/kvs: increase isolation in cache_entry_set_*
Problem: cache_entry_set_raw() and cache_entry_set_treeobj()
take ownership of a data/json_t pointer allocated in the
caller.  The caller must be trusted not to free or modify
the data afterward, or the cache will be corrupted.

Copy the data, making it easier to guarantee that the result
is not modified or freed by the original caller. In the
case of json_t, cache the encoded data but skip caching
the original object.  The object will be decoded from the
data (and then cached) upon first access.

Update users in kvs.c and commit.c.

Update unit tests.