t/kvs: Remove json type tests
Remove tests related to the json type stored in the KVS.  The KVS
now stores data in raw format, and many of the original kvs_get_type (i.e.
kvs_get_double, kvs_get_int64) are now gone.  So the tests no longer
seem necessary.  The user is simply responsible for storing the data
in the KVS they want.

As a consequence, the test binaries kvs/basic and kvs/getas are no longer
used, so remove those files.