modules/kvs: fix confusing root naming
Problem: the name "rootdir" is used to refer to the root
blobref, which is easily confused with the root directory
object.

Rename "rootdir" to "rootref" where appropriate.

Create a 'struct kvsroot' to hold the root blobref and
sequence number, since these are tightly coupled and
there will need to be more than one pair for when
planned support for multiple namespaces is implemented.