modules/kvs: setroot event handler cleanup
Problem: the important kvs.setroot event handler is a
bit complicated by the fact that the directory object is
optionally provided in the event message as an optimization.

Split the optimization off to a separate function,
don't try too hard to handle the case of an existing but
dirty/invalid cache entry, and recompute the hash after
encoding the directory rather than trusting the hash received
in the message to match.

This eliminates the last user of cache_entry_set_treeobj().