build: check for yaml-cpp usability during configure
On some distributions, yaml-cpp is packaged without a dependency
on libboost-dev, and therefore may result in compile-time errors
not detected in flux-core configure.

Do a simple preprocessor check to make sure yaml-cpp/yaml.h can
be included when YAMLCPP support is found, and issue a fatal error
if not (with a message to check config.log).

Fixes #1280